+++ app/services/script_generator/process_brainrot_character.rb
@@ -1,31 +1,13 @@
 module ScriptGenerator
-  class ProcessBrainrotCharacter
-    include Memery
-
-    def self.call(...)
-      new(...).call
-    end
-
-    def initialize(chat_id:)
-      @chat_id = chat_id
-    end
-
-    def call
-      script_processor.call(script: character_description)
-    end
-
+  class ProcessBrainrotCharacter < ProcessCharacterBase
     private
 
-    attr_reader :chat_id
-
-    delegate :character_description, to: :brainrot_character_context
-
-    memoize def brainrot_character_context
-      ScriptGenerator::BrainrotCharacterContext.new(chat_id:)
+    def character_context_class
+      ScriptGenerator::BrainrotCharacterContext
     end
 
-    memoize def script_processor
-      ScriptGenerator::ProcessScript.new(chat_id:, category: ContentCategory::BRAINROT_CHARACTER)
+    def script_category
+      ContentCategory::BRAINROT_CHARACTER
     end
   end
 end
